refactor: reorganize node modules with clear selector/mixer/viz separation
- Merge band_selection.py + selector.py into channel_selector.py
- Absorb concrete_selector.py into channel_mixer.py (rename ConcreteBandSelector → ConcreteChannelMixer)
- Rename pca.py to dimensionality_reduction.py
- Split visualizations.py into anomaly_visualization.py and pipeline_visualization.py
- Rename DRCNNTensorBoardViz → PipelineComparisonVisualizer
- Rename 8 band_selection classes (Band→Channel, FalseRGB→descriptive)
- Update __init__.py exports with all node classes
- Delete 6 old files
