+++ cuvis_ai/node/__init__.py
@@ -1,34 +1,50 @@
 """Convenience exports for node base classes and marker mixins."""
 
-from cuvis_ai.node.band_selection import (
-    BandSelectorBase,
-    BaselineFalseRGBSelector,
-    CIRFalseColorSelector,
-    HighContrastBandSelector,
-    SupervisedBandSelectorBase,
-    SupervisedCIRBandSelector,
-    SupervisedFullSpectrumBandSelector,
-    SupervisedWindowedFalseRGBSelector,
+from cuvis_ai.node.anomaly_visualization import AnomalyMask, RGBAnomalyMask, ScoreHeatmapVisualizer
+from cuvis_ai.node.channel_mixer import ConcreteChannelMixer, LearnableChannelMixer
+from cuvis_ai.node.channel_selector import (
+    ChannelSelectorBase,
+    CIRSelector,
+    FixedWavelengthSelector,
+    HighContrastSelector,
+    SoftChannelSelector,
+    SupervisedCIRSelector,
+    SupervisedFullSpectrumSelector,
+    SupervisedSelectorBase,
+    SupervisedWindowedSelector,
+    TopKIndices,
 )
+from cuvis_ai.node.dimensionality_reduction import TrainablePCA
 from cuvis_ai.node.labels import BinaryAnomalyLabelMapper
 from cuvis_ai.node.normalization import IdentityNormalizer, MinMaxNormalizer, SigmoidNormalizer
-from cuvis_ai.node.pca import TrainablePCA
-from cuvis_ai.node.selector import SoftChannelSelector, TopKIndices
+from cuvis_ai.node.pipeline_visualization import (
+    CubeRGBVisualizer,
+    PCAVisualization,
+    PipelineComparisonVisualizer,
+)
 
 __all__ = [
-    "BandSelectorBase",
-    "BaselineFalseRGBSelector",
+    "AnomalyMask",
     "BinaryAnomalyLabelMapper",
-    "CIRFalseColorSelector",
-    "HighContrastBandSelector",
+    "ChannelSelectorBase",
+    "CIRSelector",
+    "ConcreteChannelMixer",
+    "CubeRGBVisualizer",
+    "FixedWavelengthSelector",
+    "HighContrastSelector",
     "IdentityNormalizer",
+    "LearnableChannelMixer",
     "MinMaxNormalizer",
+    "PCAVisualization",
+    "PipelineComparisonVisualizer",
+    "RGBAnomalyMask",
+    "ScoreHeatmapVisualizer",
     "SigmoidNormalizer",
     "SoftChannelSelector",
-    "SupervisedBandSelectorBase",
-    "SupervisedCIRBandSelector",
-    "SupervisedFullSpectrumBandSelector",
-    "SupervisedWindowedFalseRGBSelector",
+    "SupervisedCIRSelector",
+    "SupervisedFullSpectrumSelector",
+    "SupervisedSelectorBase",
+    "SupervisedWindowedSelector",
     "TopKIndices",
     "TrainablePCA",
 ]
